// src/types/Products.ts
export type Product = {
  image: any;
  id: number;
  sku: string;
  meta: object;
  tags: string[];
  brand: string;
  title: string;
  price: number;
  stock: number;
  images: string[];
  rating: number;
  reviews: object[];
  category: string;
  quantity: number;
  thumbnail: string;
  dimensions: object;
  description: string;
  returnPolicy: string;
  installments: string;
  availabilityStatus: string;
  discountPercentage: number;
  warrantyInformation: string;
  shippingInformation: string;
  minimumOrderQuantity: number;
};